V6.2.2 - Wand/Blaster Fixes (#37)
* Consolidate Mode Original checks

Creates new shared library check for determining if pack is active or not when in Mode Original

* Update blinking light functionality

Updates the blinking light on the top of the wand body to be a barrel safety indicator in Mode Super Hero for 84/89 to better match the on-screen prop behavior, disables this light when in error states, and adjusts checks for the power-on reminder blinking light on the wand and blaster

* Fix web UI music selection on wand/blaster

Fixes issue where selecting a music track via the web UI was completely broken on the wand and blaster

* Fix missing power-on reminders

Fixes missing power-on reminder arm/disarm when controlling music on a standalone wand or the blaster from the web UI rather than the physical controls

+++ source/SingleShot/include/Header.h
@@ -366,7 +366,7 @@ bool b_device_boot_error_on = false;
  * A timer to turn on some Single-Shot Blaster lights when the system is shut down after some inactivity, as a reminder you left your power on to the system.
  */
 millisDelay ms_power_indicator;
-const uint32_t i_ms_power_indicator = 60000; // 1 minute -> 60000 milliseconds
+const uint16_t i_ms_power_indicator = 60000; // 1 minute -> 60000 milliseconds
 const uint16_t i_ms_power_indicator_blink = 500;
 
 /*
